SQOOP-1764: Numeric Overflow when getting extent map (David Robson via Venkat Ranganathan)

+++ src/java/org/apache/sqoop/manager/oracle/OraOopOracleDataChunk.java
@@ -29,7 +29,7 @@ public abstract class OraOopOracleDataChunk implements Writable {
 
   private String id;
 
-  public abstract int getNumberOfBlocks();
+  public abstract long getNumberOfBlocks();
 
   public String getWhereClause() {
     return "1=1";
